modularize rgb tests
