+++ src/rgb.rs
@@ -8,6 +8,7 @@ use rgbstd::{
 use strict_encoding::StrictSerialize;
 
 pub mod accept;
+pub mod commits;
 pub mod constants;
 pub mod invoice;
 pub mod issue;
@@ -19,10 +20,10 @@ pub mod structs;
 use crate::{
     constants::BITCOIN_ELECTRUM_API,
     rgb::{
+        commits::extract_psbt_commit,
         invoice::{accept_payment, create_invoice as create_rgb_invoice, pay_invoice},
         issue::issue_contract as create_contract,
         psbt::create_psbt as create_rgb_psbt,
-        psbt::extract_commit,
         resolvers::ExplorerResolver,
     },
     structs::{
@@ -132,7 +133,7 @@ pub async fn pay_asset(request: RgbTransferRequest) -> Result<RgbTransferRespons
     let mut stock = Stock::default();
     let (psbt, transfer) = pay_invoice(rgb_invoice, psbt, &mut stock)?;
 
-    let commit = extract_commit(psbt.clone())?;
+    let commit = extract_psbt_commit(psbt.clone())?;
     let psbt = psbt.to_string();
     let consig = RgbTransferResponse {
         consig_id: transfer.bindle_id().to_string(),
@@ -157,7 +158,7 @@ pub async fn accept_transfer(request: AcceptRequest) -> Result<AcceptResponse> {
     let mut tx_resolver = ExplorerResolver {
         explorer_url: explorer_url.to_string(),
     };
-    let resp = match accept_payment(consignment, true, &mut tx_resolver, &mut stock) {
+    let resp = match accept_payment(consignment, false, &mut tx_resolver, &mut stock) {
         Ok(transfer) => AcceptResponse {
             contract_id: transfer.contract_id().to_string(),
             transfer_id: transfer.transfer_id().to_string(),
