+++ src/rgb/selection.rs
@@ -1,0 +1,1 @@
+pub fn calibration() {}
