--- tests/lib.rs
@@ -1,2 +1,0 @@
-#[macro_use]
-pub mod rgb_test_utils;
