+++ tests/rgb.rs
@@ -1,0 +1,14 @@
+mod rgb {
+    mod utils;
+    mod unit {
+        mod commits;
+        mod invoice;
+        mod issue;
+        mod psbt;
+        mod stock;
+    }
+
+    mod integration {
+        mod selection;
+    }
+}
