+++ tests/rgb/unit/issue.rs
@@ -3,8 +3,7 @@ use anyhow::Result;
 use bitmask_core::{rgb::issue::issue_contract, util::init_logging};
 use rgbstd::persistence::Stock;
 
-mod rgb_test_utils;
-use rgb_test_utils::DumbResolve;
+use crate::rgb::utils::DumbResolve;
 
 #[tokio::test]
 async fn issue_contract_test() -> Result<()> {
